+++ engine/server/engine/centralized_logs/loki_logs_database_client.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/gorilla/websocket"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/logs_database_functions/implementations/loki"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/logs_database_functions/implementations/loki/tags"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_key_consts"
@@ -22,9 +23,11 @@ import (
 const (
 	defaultHttpClientTimeOut = 1 * time.Minute
 
-	httpProtocol              = "http://"
+	httpScheme                = "http"
+	websocketScheme           = "ws"
 	baseLokiApiPath           = "/loki/api/v1"
 	queryRangeEndpointSubpath = "/query_range"
+	tailEndpointSubpath       = "/tail"
 
 	//the global retention period store logs for 30 days = 720h.
 	maxRetentionPeriodHours = loki.LimitsRetentionPeriodHours * time.Hour
@@ -34,18 +37,24 @@ const (
 	//tenantID = enclaveID
 	organizationIdHttpHeaderKey = "X-Scope-OrgID"
 
-	startTimeQueryParamKey = "start"
-	queryLogsQueryParamKey = "query"
+	startTimeQueryParamKey    = "start"
+	queryLogsQueryParamKey    = "query"
 	entriesLimitQueryParamKey = "limit"
-	directionQueryParamKey = "direction"
+	directionQueryParamKey    = "direction"
+	delayForQueryParamKey     = "delay_for"
 
 	//We are establishing a big limit in order to get all the user-service's logs in one request
 	//We will improve this in the future generating a pagination mechanism based on the limit number
 	//and the unix epoch time (as the start time for the next request) returned by newest stream's log line
 	//We chose 4k because 4000 x 1kb lines = 4 MB, which is just under the Protobuf limit of 5MB
 	defaultEntriesLimit = "4000"
+	//For tailing logs a lower entries limit is established because the websocket endpoint will be constantly streaming the new lines
+	defaultEntriesLimitForTailingLogs = "100"
 	//The oldest item is first when using direction=forward
 	defaultDirection = "forward"
+
+	//The number of seconds to delay retrieving logs to let slow loggers catch up. Defaults to 0 and cannot be larger than 5.
+	defaultDelayForSeconds = "0"
 
 	disjunctionTagOperator = "|"
 
@@ -55,6 +64,13 @@ const (
 
 	streamValueLogLineIndex = 1
 
+	//Left the connection open from the server-side for 4 days
+	maxAllowedWebsocketConnectionDurationOnServerSide = loki.TailMaxDurationHours * time.Hour
+
+	oneHourLess = -time.Hour
+
+	logsByKurtosisUserServiceGuidChanBuffSize = 5
+	errorChanBuffSize                         = 2
 )
 
 // A backoff schedule for when and how often to retry failed HTTP
@@ -72,24 +88,29 @@ var httpRetriesBackoffSchedule = []time.Duration{
 type lokiQueryRangeResponse struct {
 	Status string `json:"status"`
 	Data   *struct {
-		ResultType string `json:"resultType"`
-		Result     []struct {
-			Stream struct {
-				KurtosisContainerType string `json:"comKurtosistechContainerType"`
-				KurtosisGUID          string `json:"comKurtosistechGuid"`
-			} `json:"stream"`
-			Values [][]string `json:"values"`
-		} `json:"result"`
+		ResultType string            `json:"resultType"`
+		Result     []lokiStreamValue `json:"result"`
 	} `json:"data"`
 }
 
-type logLine struct {
+type lokiStreamLogsResponse struct {
+	Streams []lokiStreamValue `json:"streams"`
+}
+
+type LokiLogLine struct {
 	Log string `json:"log"`
 }
 
 type lokiLogsDatabaseClient struct {
 	logsDatabaseAddress string
 	httpClient          httpClient
+}
+type lokiStreamValue struct {
+	Stream struct {
+		KurtosisContainerType string `json:"comKurtosistechContainerType"`
+		KurtosisGUID          string `json:"comKurtosistechGuid"`
+	} `json:"stream"`
+	Values [][]string `json:"values"`
 }
 
 func NewLokiLogsDatabaseClient(logsDatabaseAddress string, httpClient httpClient) *lokiLogsDatabaseClient {
@@ -104,33 +125,29 @@ func NewLokiLogsDatabaseClientWithDefaultHttpClient(logsDatabaseAddress string) 
 func (client *lokiLogsDatabaseClient) GetUserServiceLogs(
 	ctx context.Context,
 	enclaveID enclave.EnclaveID,
-	userServiceGuids map[service.ServiceGUID]bool,
-) (map[service.ServiceGUID][]string, error) {
-
-	resultLogsByKurtosisUserServiceGuid := map[service.ServiceGUID][]string{}
+	userServiceGUIDs map[service.ServiceGUID]bool,
+) (
+	map[service.ServiceGUID][]LogLine,
+	error,
+) {
 
 	httpHeaderWithTenantID := http.Header{}
 	httpHeaderWithTenantID.Add(organizationIdHttpHeaderKey, string(enclaveID))
 
-	kurtosisGuid := []string{}
-	for userServiceGuid := range userServiceGuids {
-		kurtosisGuid = append(kurtosisGuid, string(userServiceGuid))
+	kurtosisGuids := []string{}
+	for userServiceGuid := range userServiceGUIDs {
+		kurtosisGuids = append(kurtosisGuids, string(userServiceGuid))
 	}
 
 	maxRetentionLogsTimeParamValue := getMaxRetentionLogsTimeParamValue()
 
 	userServiceContainerTypeDockerValue := label_value_consts.UserServiceContainerTypeDockerLabelValue.GetString()
 
-	queryParamValue := getQueryParamValue(userServiceContainerTypeDockerValue, kurtosisGuid)
+	queryParamValue := getQueryParamValue(userServiceContainerTypeDockerValue, kurtosisGuids)
 
 	getLogsPath := baseLokiApiPath + queryRangeEndpointSubpath
 
-	queryRangeEndpointUrlStr := fmt.Sprintf("%v%v%v", httpProtocol, client.logsDatabaseAddress, getLogsPath)
-
-	queryRangeEndpointUrl, err := url.Parse(queryRangeEndpointUrlStr)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred parsing url string '%v'", queryRangeEndpointUrlStr)
-	}
+	queryRangeEndpointUrl := &url.URL{Scheme: httpScheme, Host: client.logsDatabaseAddress, Path: getLogsPath}
 
 	queryRangeEndpointQuery := queryRangeEndpointUrl.Query()
 
@@ -162,28 +179,135 @@ func (client *lokiLogsDatabaseClient) GetUserServiceLogs(
 		return nil, stacktrace.Propagate(err, "The response body's schema payload received '%+v' by calling the Loki's query range endpoint is not what was expected; this is a bug in Kurtosis", lokiQueryRangeResponseObj)
 	}
 
-	for _, queryRangeResult := range lokiQueryRangeResponseObj.Data.Result {
-		resultKurtosisGuidStr := queryRangeResult.Stream.KurtosisGUID
-		resultKurtosisGuid := service.ServiceGUID(resultKurtosisGuidStr)
-		resultKurtosisGuidLogLinesStr := make([]string, len(queryRangeResult.Values))
-		for queryRangeIndex, queryRangeValue := range queryRangeResult.Values {
-			logLineStr, err := getLogLineStrFromStreamValue(queryRangeValue)
-			if err != nil {
-				return nil, stacktrace.Propagate(err, "An error occurred getting log line string from stream value '%+v'", queryRangeValue)
-			}
-			resultKurtosisGuidLogLinesStr[queryRangeIndex] = logLineStr
-		}
-		resultLogsByKurtosisUserServiceGuid[resultKurtosisGuid] = resultKurtosisGuidLogLinesStr
+	lokiStreams := lokiQueryRangeResponseObj.Data.Result
+
+	resultLogsByKurtosisUserServiceGuid, err := newUserServiceLogLinesByUserServiceGuidFromLokiStreams(lokiStreams)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting user service log lines from loki streams '%+v'", lokiStreams)
 	}
 
-
 	return resultLogsByKurtosisUserServiceGuid, nil
+}
+
+func (client *lokiLogsDatabaseClient) StreamUserServiceLogs(
+	ctx context.Context,
+	enclaveID enclave.EnclaveID,
+	userServiceGUIDs map[service.ServiceGUID]bool,
+) (
+	chan map[service.ServiceGUID][]LogLine,
+	chan error,
+	func(),
+	error,
+) {
+
+	websocketDeadlineTime := getWebsocketDeadlineTime()
+
+	ctxWithDeadline, cancelCtxFunc := context.WithDeadline(ctx, websocketDeadlineTime)
+	shouldCancelCtx := true
+	defer func() {
+		if shouldCancelCtx {
+			cancelCtxFunc()
+		}
+	}()
+
+	tailLogsEndpointURL, httpHeaderWithTenantID := client.getTailLogEndpointURLAndHeader(enclaveID, userServiceGUIDs)
+
+	//this channel will return the user service log lines by service GUI
+	logsByKurtosisUserServiceGuidChan := make(chan map[service.ServiceGUID][]LogLine, logsByKurtosisUserServiceGuidChanBuffSize)
+
+	//this channel return an error if the stream fails at some point
+	streamErrChan := make(chan error, errorChanBuffSize)
+
+	tailLogsWebsocketConn, tailLogsHttpResponse, err := websocket.DefaultDialer.DialContext(ctxWithDeadline, tailLogsEndpointURL.String(), httpHeaderWithTenantID)
+	if err != nil {
+		errMsg := fmt.Sprintf("An error occurred calling the logs-database-tail-logs-endpoint with URL '%v' using headers '%+v'", tailLogsEndpointURL.String(), httpHeaderWithTenantID)
+		if tailLogsHttpResponse != nil {
+			errMsg = fmt.Sprintf("%v; the '%v' status code was received from the server", errMsg, tailLogsHttpResponse.StatusCode)
+		}
+		return nil, nil, nil, stacktrace.Propagate(err, errMsg)
+	}
+
+	go runReadStreamResponseAndAddUserServiceLogLinesToUserServiceLogsChannel(
+		ctx,
+		cancelCtxFunc,
+		tailLogsWebsocketConn,
+		logsByKurtosisUserServiceGuidChan,
+		streamErrChan,
+	)
+
+	//We need to cancel the websocket connection only if something fails because we need the connection open after returning
+	shouldCancelCtx = false
+	return logsByKurtosisUserServiceGuidChan, streamErrChan, cancelCtxFunc, nil
 }
 
 // ====================================================================================================
 //                                       Private helper functions
 // ====================================================================================================
-func (client *lokiLogsDatabaseClient) doHttpRequestWithRetries(request *http.Request) ([]byte, error)  {
+func runReadStreamResponseAndAddUserServiceLogLinesToUserServiceLogsChannel(
+	ctx context.Context,
+	cancelCtxFunc context.CancelFunc,
+	tailLogsWebsocketConn *websocket.Conn,
+	logsByKurtosisUserServiceGuidChan chan map[service.ServiceGUID][]LogLine,
+	errChan chan error,
+) {
+
+	//Closing all the open resources at the end
+	defer func() {
+		if err := tailLogsWebsocketConn.Close(); err != nil {
+			logrus.Warnf("We tried to close the tail logs websocket connection, but doing so threw an error:\n%v", err)
+		}
+		close(logsByKurtosisUserServiceGuidChan)
+		close(errChan)
+		cancelCtxFunc()
+	}()
+
+	for {
+
+		streamResponse := &lokiStreamLogsResponse{}
+
+		if readingStreamResponseErr := tailLogsWebsocketConn.ReadJSON(streamResponse); readingStreamResponseErr != nil {
+
+			logrus.Debugf("Reading the tail logs streams response has returned the following error:\n'%v'", readingStreamResponseErr)
+
+			if websocket.IsCloseError(readingStreamResponseErr) {
+				logrus.Debug("Reading the tail logs streams has been closed")
+				return
+			}
+
+			if ctxErr := ctx.Err(); ctxErr != nil {
+				switch ctx.Err() {
+				case context.Canceled:
+					logrus.Debug("The tail logs streams context has been canceled")
+					return
+				case context.DeadlineExceeded:
+					logrus.Debug("The tail logs streams context deadline has been exceeded")
+					deadlineErrMsg := "Reading the tail logs streams has exceeded the deadline time"
+					ctxDeadlineTime, ok := ctx.Deadline()
+					if ok {
+						deadlineErrMsg = fmt.Sprintf("%v with value '%v'", deadlineErrMsg, ctxDeadlineTime)
+					}
+					errChan <- stacktrace.NewError(deadlineErrMsg)
+					return
+				default:
+					logrus.Debugf("The tail logs streams context contains this error '%v' ", ctxErr)
+				}
+			}
+
+			errChan <- stacktrace.Propagate(readingStreamResponseErr, "An error occurred reading the Loki's tail log endpoint")
+			return
+		}
+
+		//Does the reading
+		resultLogsByKurtosisUserServiceGuid, err := newUserServiceLogLinesByUserServiceGuidFromLokiStreams(streamResponse.Streams)
+		if err != nil {
+			errChan <- stacktrace.Propagate(err, "An error occurred getting user service log lines from loki streams '%+v'", streamResponse.Streams)
+			return
+		}
+		logsByKurtosisUserServiceGuidChan <- resultLogsByKurtosisUserServiceGuid
+	}
+}
+
+func (client *lokiLogsDatabaseClient) doHttpRequestWithRetries(request *http.Request) ([]byte, error) {
 
 	var (
 		httpResponse          *http.Response
@@ -221,18 +345,17 @@ func (client *lokiLogsDatabaseClient) doHttpRequestWithRetries(request *http.Req
 	return nil, stacktrace.NewError("The request '%+v' could not be executed successfully, even after applying this retry backoff schedule '%+v', the status code '%v' and body '%v' were the last one received", request, httpRetriesBackoffSchedule, httpResponse.StatusCode, string(httpResponseBodyBytes))
 }
 
-
 func (client *lokiLogsDatabaseClient) doHttpRequest(
 	request *http.Request,
 ) (
 	resultResponse *http.Response,
 	resultResponseBodyBytes []byte,
 	resultErr error,
-)  {
+) {
 
 	var (
 		httpResponseBodyBytes []byte
-		err error
+		err                   error
 	)
 
 	httpResponse, err := client.httpClient.Do(request)
@@ -291,18 +414,98 @@ func getKurtosisGuidParamValues(kurtosisGuids []string) string {
 	return kurtosisGuidsParamValues
 }
 
-func getLogLineStrFromStreamValue(streamValue []string) (string, error) {
+func (client *lokiLogsDatabaseClient) getTailLogEndpointURLAndHeader(
+	enclaveID enclave.EnclaveID,
+	userServiceGuids map[service.ServiceGUID]bool,
+) (url.URL, http.Header) {
+
+	kurtosisGuids := []string{}
+	for userServiceGuid := range userServiceGuids {
+		kurtosisGuids = append(kurtosisGuids, string(userServiceGuid))
+	}
+
+	maxRetentionLogsTimeForTailingLogsParamValue := getStartTimeForStreamingLogsParamValue()
+
+	userServiceContainerTypeDockerValue := label_value_consts.UserServiceContainerTypeDockerLabelValue.GetString()
+
+	queryParamValue := getQueryParamValue(userServiceContainerTypeDockerValue, kurtosisGuids)
+
+	tailLogsPath := baseLokiApiPath + tailEndpointSubpath
+
+	tailLogsEndpointUrl := url.URL{Scheme: websocketScheme, Host: client.logsDatabaseAddress, Path: tailLogsPath}
+
+	tailLogsEndpointQuery := tailLogsEndpointUrl.Query()
+
+	tailLogsEndpointQuery.Set(queryLogsQueryParamKey, queryParamValue)
+	tailLogsEndpointQuery.Set(delayForQueryParamKey, defaultDelayForSeconds)
+	tailLogsEndpointQuery.Set(entriesLimitQueryParamKey, defaultEntriesLimitForTailingLogs)
+	tailLogsEndpointQuery.Set(startTimeQueryParamKey, maxRetentionLogsTimeForTailingLogsParamValue)
+
+	tailLogsEndpointUrl.RawQuery = tailLogsEndpointQuery.Encode()
+
+	httpHeaderWithTenantID := http.Header{}
+	httpHeaderWithTenantID.Add(organizationIdHttpHeaderKey, string(enclaveID))
+
+	return tailLogsEndpointUrl, httpHeaderWithTenantID
+}
+
+func getStartTimeForStreamingLogsParamValue() string {
+	now := time.Now()
+	startTime := now.Add(oneHourLess)
+	startTimeNano := startTime.UnixNano()
+	startTimeNanoStr := fmt.Sprintf("%v", startTimeNano)
+	return startTimeNanoStr
+}
+
+func newUserServiceLogLinesByUserServiceGuidFromLokiStreams(lokiStreams []lokiStreamValue) (map[service.ServiceGUID][]LogLine, error) {
+
+	resultLogsByKurtosisUserServiceGuid := map[service.ServiceGUID][]LogLine{}
+
+	for _, queryRangeResult := range lokiStreams {
+		resultKurtosisGuidStr := queryRangeResult.Stream.KurtosisGUID
+		resultKurtosisGuid := service.ServiceGUID(resultKurtosisGuidStr)
+		resultKurtosisGuidLogLines := make([]LogLine, len(queryRangeResult.Values))
+		for queryRangeIndex, queryRangeValue := range queryRangeResult.Values {
+			logLineObj, err := newLogLineFromStreamValue(queryRangeValue)
+			if err != nil {
+				return nil, stacktrace.Propagate(err, "An error occurred getting log line string from stream value '%+v'", queryRangeValue)
+			}
+			resultKurtosisGuidLogLines[queryRangeIndex] = *logLineObj
+		}
+
+		userServiceLogLines, found := resultLogsByKurtosisUserServiceGuid[resultKurtosisGuid]
+		if found {
+			userServiceLogLines = append(userServiceLogLines, resultKurtosisGuidLogLines...)
+		} else {
+			userServiceLogLines = resultKurtosisGuidLogLines
+		}
+
+		resultLogsByKurtosisUserServiceGuid[resultKurtosisGuid] = userServiceLogLines
+	}
+
+	return resultLogsByKurtosisUserServiceGuid, nil
+}
+
+func newLogLineFromStreamValue(streamValue []string) (*LogLine, error) {
 	if len(streamValue) > streamValueNumOfItems {
-		return "", stacktrace.NewError("The stream value '%+v' should contains only 2 items but '%v' items were found, this should never happen; this is a bug in Kurtosis", streamValue, len(streamValue))
+		return nil, stacktrace.NewError("The stream value '%+v' should contains only 2 items but '%v' items were found, this should never happen; this is a bug in Kurtosis", streamValue, len(streamValue))
 	}
 
-	logLineStr := streamValue[streamValueLogLineIndex]
-	logLineBytes := []byte(logLineStr)
-	logLineObj := &logLine{}
+	lokiLogLineStr := streamValue[streamValueLogLineIndex]
+	lokiLogLineBytes := []byte(lokiLogLineStr)
+	lokiLogLine := &LokiLogLine{}
 
-	if err := json.Unmarshal(logLineBytes, logLineObj); err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred unmarshalling logline '%+v'", logLineObj)
+	if err := json.Unmarshal(lokiLogLineBytes, lokiLogLine); err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred unmarshalling Loki log line '%+v'", lokiLogLine)
 	}
 
-	return logLineObj.Log, nil
+	newLogLineObj := newLogLine(lokiLogLine.Log)
+
+	return newLogLineObj, nil
+}
+
+func getWebsocketDeadlineTime() time.Time {
+	now := time.Now()
+	deadlineTime := now.Add(maxAllowedWebsocketConnectionDurationOnServerSide)
+	return deadlineTime
 }