+++ docs/changelog.md
@@ -1,4 +1,12 @@
 # TBD
+### Features
+- Implemented the new `StreamUserServiceLogs` endpoint in the Kurtosis engine server
+- Added the new `StreamUserServiceLogs` in the Kurtosis engine golang library
+- Added the `StreamUserServiceLogs` method in Loki logs database client
+- Added the `StreamUserServiceLogs` method in Kurtosis backend logs client
+
+### Changes
+- Updated the CLI `service logs` command in order to use the new `KurtosisContext.StreamUserServiceLogs` when user requested to follow logs
 
 ### Changes
 - InterpretationError is now able to store a `cause`. It simplifies being more explicit on want the root issue was
@@ -20,7 +28,6 @@
 - Add more integration tests for Kurtosis modules with input and output types
 
 # 0.51.7
-
 ### Fixes
 - Fixed instruction position to work with nested functions
 
@@ -28,7 +35,6 @@
 - Instruction position now contains the filename too
 
 # 0.51.6
-
 ### Features
 - Added an `import_types` Starlark instruction to read types from a .proto file inside a module
 - Added the `time` module for Starlark to the interpreter